refactor(eval): split eval-graders.js into io + per-type modules
EVAL-6: decomposed the 1134-line eval-graders.js into a cohesive shared-I/O leaf
(eval-grader-io.js) + per-type grader modules (code/model/behavioral) + a thin dispatcher.
gradeTrialResult/gradeWithMixed/gradeWithHuman stay in eval-graders.js (genuinely
cross-cutting). All 29 functions moved verbatim (byte-identical); acyclic imports
(io leaf <- type modules <- dispatcher); dispatcher re-exports the 21 public names so no
importer changed. eval-graders.js 1134 -> 230; new modules io 350 / model 340 /
behavioral 182 / code 161 (all under the 400 soft limit). Full npm test green; lint clean.

Previously deferred as risky; the cohesive io-module approach (not a per-type 'common dump')
made a clean split possible after the user prioritized it.
